Remove orange flower workflow and update output location
The `azure-static-web-apps-orange-flower-0fca8700f.yml` file has been completely removed, which included jobs for building and deploying the application as well as closing pull requests.

In the `azure-static-web-apps-red-pebble-0b405880f.yml` file, the `output_location` for the build output has been changed from `"dist"` to `"dist/browser"`, indicating a change in the directory structure for the built application content.